modifying codes in this dir and running on data
Modified the codes in order to run for a fixed structure of the filesystem of weaklensing pipeline and it's output directory I've set up.
Result of running the code on many signals at once, gives output files which are stored in the same file tree exactly like the output dir of weaklens pipeline.